% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_irf.R
\name{draw_irf}
\alias{draw_irf}
\title{Draw IRF}
\usage{
draw_irf(irf, res_var, ncol_pic = 2)
}
\arguments{
\item{irf}{a returned objective from \code{IRFrf_gen}}

\item{res_var}{a string vector composed with names of response variable}

\item{ncol_pic}{multiple pictures are arranged \code{ncol_pic} coloumns}
}
\description{
draw IRF picture according to results from \code{IRFrf_gen}
}
\examples{

}
