% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_gen_panel.R
\name{dt_gen_panel}
\alias{dt_gen_panel}
\title{Transfer Panel Data}
\usage{
dt_gen_panel(regdata, indx = c("id", "year"), s = 1, plag = 1, yname = FALSE)
}
\arguments{
\item{regdata}{a data frame, either a time series or panal data}

\item{indx}{if \code{regdata} is time series, \code{indx} is NULL. If \code{regdata}
is panal data, \code{indx} is a character string vector whose length is 2, the 1st element
is coloumn name of individual id, and the 2nd element is coloumn names of time.}

\item{s}{a horizon in local projections, default is 1.}

\item{plag}{lag order of VAR.}

\item{yname}{logical, whether return the name of dependent variables.}
}
\value{
if \code{yname} is FALSE, return a data.frame include indepedent variables and
depedent variables in VAR. If \code{yname} is TRUE, returen a list, the 1st element is
data.frame, the 2nd element is names of dependent variables.
}
\description{
Transfer Panel Data
}
