% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRFrf.R
\name{IRFrf}
\alias{IRFrf}
\title{Compute IRF by random forest method}
\usage{
IRFrf(data, pmax = 5, p = NULL, s = 12, shockvar = 1, d = 1, histime = 1)
}
\arguments{
\item{pmax}{max lag order where select a lag order which has minimum OOB MSE.}

\item{s}{the horizon of IRF}

\item{shockvar}{a numeric scalor which denotes the shock variable.}

\item{d}{the size of shock}

\item{set}{up the lag order directly, and don't select from \code{pmax}.}

\item{hist}{the numerical scalor that represents the time of shock, i.e. row index
of \code{data}.}
}
\description{
Compute IRF by random forest method
}
