% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARrf.R
\name{VARrf}
\alias{VARrf}
\title{Estimate VAR by random forest}
\usage{
VARrf(data, indx = NULL, pmax = 5, p = NULL, s = 1)
}
\arguments{
\item{data}{a data.frame, and all variables is endogenous variables in VAR.}

\item{indx}{if \code{data} is time series, \code{indx} is NULL. If \code{data}
is panal data, \code{indx} is a character string vector whose length is 2, the 1st element
is coloumn name of individual id, and the 2nd element is coloumn names of time.}

\item{pmax}{a max lag order, default value is 5.}

\item{p}{a fixed lag order. When \code{p} is set up, \code{pmax} is unuseful.}

\item{s}{a horizon in local projections, default is 1. Gnerally don't change it.}
}
\value{
An list:
\enumerate{
\item{fit_rf}{An object of class \code{rfsrc, grow} from the package \code{randomForestSRC},
and it has minimum OOB MSE in 1:\code{pmax}.}
\item{p}{optimal lag order,and it has minimum OOB MSE.}
}
}
\description{
Estimate VAR by random forest
}
\examples{
library(magrittr)
B1 <- matrix(c(0.2, 0.1, 0.2, 0.3,0.3,0.4,0.2,0.1), 2)
sim_dt <- vars::VAR.sim(B1, n = 200, lag = 2,include = 'none',show.parMat = T) \%>\%
  as.data.frame()
  colnames(sim_dt) <- c('y1','y2')
# fit a VAR with max lag order equal 5
fit <- VARrf(sim_dt, pmax = 5)
# fit a VAR with fixed lag order equal 4
fit <- VARrf(sim_dt, p = 4)
# IRF
picdata <- IRFrf(data = sim_dt, pmax = 5, s = 12, shockvar = 1)
ggplot(data = picdata, aes(x = s, y = y1)) + geom_line() +
  geom_hline(yintercept = 0) + theme_bw()
# Generalized IRF
IRFrf_gen(data = sim_dt[1:20,], pmax = 3, s = 5, shockvar = 1, ncores = 7)


}
