% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARrf_forcast.R
\name{VARrf_forcast}
\alias{VARrf_forcast}
\title{Forcast Y Using Random Forest method}
\usage{
VARrf_forcast(fit, s = 12, startvalue = NULL, shockvar = 1, d = 1, n_eps = 1)
}
\arguments{
\item{fit}{a object from \code{rfsrc}}

\item{s}{a numerical scalor, horizon of predicting.}

\item{startvalue}{a data frame with 1 row, and its coloums is same with xvar in
the estimated result from \code{rfsrc}.}

\item{shockvar}{which variable is shock variable.}

\item{d}{size of shock, default value is 1.}

\item{n_eps}{}
}
\value{
a list with 2 elements, the 1st elements is a forcast with no shock, and
the 2nd element is a forcast with shock.
}
\description{
Forcast Y Using Random Forest method
}
\details{

}
\note{
de-mean each variable
}
