% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARrf.R
\name{VARrf}
\alias{VARrf}
\title{Estimate VAR by random forest}
\usage{
VARrf(data, pmax = 5, p = NULL, s = 1)
}
\arguments{
\item{data}{a data.frame}

\item{pmax}{a max lag order}

\item{p}{a fixed lag order. When \code{p} is set up, \code{pmax} is unuseful.}

\item{s}{a horizon in local projections, default is 1. Gnerally don't change it.}
}
\value{
An list:
\enumerate{
\item{fit_rf}{An object of class \code{rfsrc, grow} from the package \code{randomForestSRC},
and it has minimum OOB MSE in 1:\code{pmax}.}
\item{p}{optimal lag order,and it has minimum OOB MSE.}
}
}
\description{
Estimate VAR by random forest
}
